requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
langchain
langgraph
langchain-google-genai
pydantic